simplest example rendering list of ideas
I put this in the following folder:

  examples/list-of-ideas

But the files have very generic names:

  examples/list-of-ideas/index.html
  examples/list-of-ideas/application.js

The idea is that to build on this example just duplicate
the whole examples/list-of-ideas folder and rename it.

I used a separate application.js file so an syntax highlighters
or checkers would have the chance to work better because the
file isn't a mixture of html AND JavaScript.

The functionality isn't any different that what's in 
./Visualization.html -- only the the structure is simpler
and the JavaScript itself is simpler because I removed
anything that wasn't specifically needed.