new example: chart of ideas over time
Very simple example using d3.js Time scales and axis.

Ideas are represented by small blue circles and are placed 
in the correct position along the X (time) axis.

Ideas are randomly distributed along the Y axis.

When you hover the mouse over and idea circle the color 
of the outline line stroke highlights from blue to orange.

Clicking on a idea circle brings up a tooltip with the 
date and idea text.