Y axis is time of day when idea was created
Right now d3.js doesn't support a time axis which
is an abstract 24 hours -- d3.js date axes and scales
expect actual JavaScript date objects.

So this time axis is a linear scale expecting time of day
in a decimal 24 hour format.

There is a special yTickTimeOfDayFormatter which takes
decimal hours as an argument and returns a string consisting
of: HH:MM:SS