refactor management of selected_data
Now the function updateSelectedData() will set the 
current_workgroup_key (either a workgroup or the variable
all_workgroups_key).

In addition moved the combination of updating and rendering
to a new function: updateSelectedDataAndRedraw();

This allows updateSelectedData() to be used early in the 
processCSVRows function before rendering is possible.

A new function: filterWorkgroupData(workgroup_key)
will return the selected set of objects without setting
selected_data. This will be useful in a subsequent commit
that supports displaying all the ideas for a workgroup in 
a tooltip.