tooltips now show all ideas for workgroup
Mouseover an idea circle to see all ideas for that workgroup
in the tooltip. Because the list might be long and a user
might need to scroll down to see all of it leave the tooltip
visible even after the mouse leaves the idea circle.

A further mousedown anywhere in the chart NOT in an idea
circle will erase the tooltip.

In addition the specific idea for that idea circle is 
displayed in the tooltip in bold. The specific styling
is set in the css -- the style could just as easily be a 
different color font … etc.